Renamed cli file, and moved hidden commands into 'auth' and 'admin' subcommands
