+++ setup.py
@@ -9,7 +9,7 @@ setuptools.setup(
     long_description=open("README.md").read(),
     long_description_content_type="text/markdown",
     packages=setuptools.find_packages(),
-    entry_points={"console_scripts": "onyx = onyx.cli2:main"},
+    entry_points={"console_scripts": "onyx = onyx.cli:main"},
     install_requires=[
         "requests",
         "typer>=0.6.0",
